pyproject.toml
[tool.poetry]
name = "wcg-pdf-to-csv"
version = "0.1.0"
description = ""
authors = ["pandelvgun <dgundani2015@mysbf.co.za>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
camelot-py = "^1.0.0"
tcl = "^0.2"
pymupdf = "^1.25.5"
ghostscript = "^0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
